[finance_report_audit] refactor(money): consolidate lib/currency into lib/money, retire currency.ts (#1167)
Fold the frontend lib/currency.ts (display/formatting + Decimal helpers) into the
lib/money module so the frontend has ONE money module, and delete currency.ts.
Behaviour-preserving move:

- lib/currency.ts -> lib/money/format.ts (verbatim), re-exported from lib/money.
- lib/currency.test.ts -> lib/money/format.test.ts.
- Repoint all 37 importers (@/lib/currency -> @/lib/money; portfolioPerformance
  ./currency -> ./money). currency.ts deleted.
- Drop the stale traceability-exceptions entry for the moved test (format.test.ts
  already references AC2.8.2 — Decimal Safety).

Note on rounding: the only HALF_UP in the old module was *display* formatting
(toFixed/toDecimalPlaces), which is conventional for display and is intentionally
preserved (no UX/snapshot change). Money *value* rounding (2 dp, banker's
HALF_EVEN) lives in the lib/money value types and is conformance-locked to the
backend via common/money/conformance/vectors.json — frontend value computation
here is exact Decimal, so there is no cross-language value divergence to fix.

Verify: full frontend vitest green (913 tests; the 4 unrelated OTel files fail
only on missing local @opentelemetry deps, pass under CI npm ci); eslint clean;
no money/currency tsc errors; doc-consistency + no-AC-classification gates green.

Refs #1167.
