+++ apps/frontend/src/app/(main)/journal/page.tsx
@@ -8,7 +8,7 @@ import ConfirmDialog from "@/components/ui/ConfirmDialog";
 import ConfidenceBadge from "@/components/ui/ConfidenceBadge";
 import { useToast } from "@/components/ui/Toast";
 import { apiFetch } from "@/lib/api";
-import { formatCurrencyLocale, sumAmounts } from "@/lib/currency";
+import { formatCurrencyLocale, sumAmounts } from "@/lib/money";
 import { formatDateDisplay } from "@/lib/date";
 import { JournalEntry, JournalEntryListResponse, JournalLine } from "@/lib/types";
