+++ apps/frontend/src/components/assets/GuidedEvidenceForm.tsx
@@ -21,7 +21,7 @@ import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { Badge } from "@/components/ui";
 import { useToast } from "@/components/ui/Toast";
 import { apiFetch } from "@/lib/api";
-import { formatCurrencyLocale } from "@/lib/currency";
+import { formatCurrencyLocale } from "@/lib/money";
 import { formatDateDisplay, formatDateInput } from "@/lib/date";
 import type {
   ManualValuationBasis,
