+++ apps/frontend/src/components/reconciliation/Workbench.tsx
@@ -7,7 +7,7 @@ import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { BackLink } from "@/components/ui/BackLink";
 import { InfoHint } from "@/components/ui/InfoHint";
 import { apiFetch } from "@/lib/api";
-import { formatAmount } from "@/lib/currency";
+import { formatAmount } from "@/lib/money";
 import {
   ReconciliationMatchListResponse,
   ReconciliationMatchResponse,
